requirements.txt
Django>=5.1
gunicorn==25.1.0
boto3==1.40.59
python-dotenv==1.1.1
llama-index
llama-index-core
llama-index-llms-openai
llama-index-retrievers-bedrock
asgiref
whitenoise==6.6.0
psycopg2-binary
